requirements.txt
Django==4.1
djangorestframework==3.14.0
dynaconf==3.1.12
